requirements.txt
# Core dependencies
pydantic>=2.0.0
chromadb>=0.4.0
openai>=1.0.0
numpy>=1.24.0
aiosqlite>=0.19.0
python-ulid>=2.0.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
hypothesis>=6.0.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.0.0

# Optional: LangChain integration
# langchain>=0.1.0
# langchain-core>=0.1.0
